Systems

Potentials

OrbitalElements.ψ — Function

ψ(r, model::Potential)

Potential value for central potential model at radius r.

the Hernquist potential

OrbitalElements.dψ — Function

dψ(r, model::Potential)

Potential derivative value for central potential model at radius r.

the Hernquist potential derivative

OrbitalElements.d2ψ — Function

d2ψ(r, model::Potential)

Potential second derivative for central potential model at radius r.

the Hernquist potential second derivative

Distribution functions

Changes of coordinates

Isochrone specific functions

Plummer specific functions

Utility functions